DUSK/requirements.txt
flask>=3.0
smbus2>=0.4
luma.oled>=3.12
luma.core>=2.4
Pillow>=10.0
pigpio>=1.78
RPi.GPIO>=0.7
picamera2>=0.3
numpy>=1.24
